Sprinkle some intptr_t
There's going to be breakage here

+++ CODE/ioobj.cpp
@@ -371,7 +371,7 @@ void FactoryClass::Decode_Pointers(void)
 		assert(Object != NULL);
 	}
 
-	((HouseClass *&)House) = HouseClass::As_Pointer((HousesType)House);
+	((HouseClass *&)House) = HouseClass::As_Pointer((HousesType)(intptr_t)House);
 	assert(House != NULL);
 }
